[console] reduce blinking while displaying diff/source
Pygments may take a while to highlight the diff/source and it induce lag while
browing the files.

The idea is to delay colorizing the text. The text was previously displayed
whitout color and then highlighted => blinking. The beggining of the text is
now highlighted immediatly which reduces the blinking effect.