[qt] goto: do not remove non validated-revset on closing the toolbar
The previous behaviour requires that the user validate the revset
by hitting [Return]. If the revset were not validated, it was removed
while closing the toolbar.

This is annoying and we were forced to click on [Return].
Instead, we expect to go on the next found revision on clicking on [Return].

As a consequence, the first revision of the revset should not be selected as
it may causes annoying big moves.