[windows] py2exe & innosetup: redefine log file location
The default log file is created in the same folder as the application.
But the default installation directory using innosetup (and perhaps other install tools)
is non-writable by user (a.k.a. c:\Pogram Files).

So we redefine the original location of the log file.
This parameter is fixed as a default argument value of the "write" method of the object that patch the original sys.stderr (by py2exe).
We have to rewrite the object class in order to tune this value.


.. note:: The module that contains the original class could not be imported because of a NameError ("missing _MessageBox": see py2exe.boot_common).
          So we use introspection (sys.stderr.__class__) to get an access to this class.